StoreOpener replace totally the RrdPool (too much lock), the back end can now be specified (only files based on)

+++ src/jrds/probe/snmp/PartitionSpace.java
@@ -60,10 +60,10 @@ public class PartitionSpace extends RdsIndexedSnmpRrd {
 				allocUnit = value.intValue();
 			}
 			else if(totalOid.equals(oid)) {
-				total = value.intValue();
+				total = value.longValue();
 			}
 			else if(usedOid.equals(oid)) {
-				used = value.intValue();
+				used = value.longValue();
 			}
 		}
 		total *= allocUnit;
